dev-requirements.txt
pygame-ce
black
invoke
pyinstaller
psutil